added bnb, twoapprox algs, added results

+++ Lab2/algorithms/optimal_result.py
@@ -6,7 +6,7 @@ class OptimalResult:
     def __init__(self, N):
         self.knapsack_cost = 0
         self.knapsack_weight = 0
-        self.item_indexes = np.zeros(N, dtype=np.uint32)
+        self.item_indexes = np.zeros(N, dtype=np.float32)
     
     def __repr__(self):
         return f"Cost of knapsack: {self.knapsack_cost} \nWeight of knapsack: {self.knapsack_weight}\nIncluded items: {self.item_indexes}"
